Add support for scripts in the TextBox widget
The previous TextBox widget allowed to include scripts in its contents so that you could use it as a means to add some interactive content in every page in a simple way. The new version doesn't allow this by default.

I've added a line to the TinyMCE HTM Editor so that it allows iframes and scripts in its contents.